542 data validator (#606)
#542 Data validator

New data validator reuses pipelines infrastructure and becomes a part of pipelines, improves maintainability, code relevance, and improves performance.

The main modules are:
- pipelines/gbif/validator
- pipelines/gbif/coordinator/crawler/../metrics
- pipelines/gbif/coordinator/crawler/../validator

The validator uses the old response model for WS and distributed tasks workflow where each task runs its own piece of work and delegates to the next task:

The tasks graph (all steps go via tasks/BalancerCallback):
validator/pipelines-validator-ws -> tasks/ArchiveValidatorCallback -> tasks/DwcaToAvroCallback -> tasks/InterpretationCallback -> tasks/IndexingCallback -> tasks/MetricsCollectorCallback (or validator/validator-checklists for checklists)

We reuse the existing ingestion workflow and add 2 more steps: tasks/ArchiveValidatorCallback and tasks/MetricsCollectorCallback

- tasks/ArchiveValidatorCallback validates a DWCA and tests all main attributes, such as correct eml, necessary fields and etc.
- tasks/MetricsCollectorCallback collect metrics about terms, occurrence issues and etc using ES index 

To control that workflow we add a flag into validatorOnly into all tasks configs and also into ingestion jobs to skip some code unnecessary for the validator

+++ tools/archives-converters/src/main/java/org/gbif/converters/converter/ConverterToVerbatim.java
@@ -29,6 +29,8 @@ public abstract class ConverterToVerbatim {
   private Path outputPath;
   private Path metaPath;
 
+  private boolean skipDeletion = false;
+
   public ConverterToVerbatim hdfsSiteConfig(String hdfsSiteConfig) {
     this.hdfsSiteConfig = hdfsSiteConfig;
     return this;
@@ -79,12 +81,17 @@ public abstract class ConverterToVerbatim {
     return this;
   }
 
+  public ConverterToVerbatim skipDeletion(boolean skipDeletion) {
+    this.skipDeletion = skipDeletion;
+    return this;
+  }
+
   public boolean convert() {
 
     Objects.requireNonNull(inputPath, "inputPath cannot be null");
     Objects.requireNonNull(outputPath, "outputPath cannot be null");
 
-    boolean isConverted;
+    boolean isConverted = false;
 
     // the fs has to be out of the try-catch block to avoid closing it, because the hdfs client
     // tries to reuse the
@@ -110,7 +117,9 @@ public abstract class ConverterToVerbatim {
       log.error("Failed performing conversion on {}", inputPath, e);
       throw new IllegalStateException("Failed performing conversion on " + inputPath, e);
     } finally {
-      isConverted = deleteAvroFileIfEmpty(fs, outputPath);
+      if (!skipDeletion) {
+        isConverted = deleteAvroFileIfEmpty(fs, outputPath);
+      }
     }
 
     return !isConverted;
